slabs,sectors: move and alphabetize all the things
